refactor: create interchange details section component
